Add index_template.html file.
It is needed because we can specify <div id="root"></div> which is needed for the React app.

index_template.html is used as a template for the index.html file in the ./dist directory which is created during building the app.

It is also used as a template for the HtmlWebpackPlugin which also creates index.html file but keeps it in memory for development purposes.